requirements.txt
python-dotenv==1.0.0
requests==2.31.0
openai==1.12.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0
Pillow==10.2.0
retry==0.9.2

